minimize skip list for current fw submodule version
As once added to the list we never check if it'd be available now we
have some unnecessary entries in the skip list, which this patch
cleans up. Produced by deleting all skip list entries and successive
re-adding them by doing:
./assemble-firmware.pl fwlist-FILE /tmp) 2>&1 | awk '/unable/ { print $5 }' >> skip-cleanedup
and adding only those which are now not found.

The "additions" come from those lines were previously multiple
skipped fws were placed on a single one.

We could also add support for such checks in the assemble-firmware
script, wouldn't be to hard, but for now do this as a starter.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>